Refactors encrypted?
